pyproject.toml
[tool.poetry]
name = "fastapi-product-service"
version = "0.1.0"
description = ""
authors = ["artorias <kerelkobarabash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
strawberry-graphql = {extras = ["fastapi"], version = "^0.229.1"}
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
punq = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
